--- packages/rocketchat-authorization/lib/rocketchat.js
@@ -1,3 +1,0 @@
-import { RocketChat } from 'meteor/rocketchat:lib';
-
-RocketChat.authz = {};
