+++ packages/rocketchat-lib/client/models/Messages.js
@@ -1,0 +1,4 @@
+import { Messages } from 'meteor/rocketchat:models';
+import _ from 'underscore';
+
+RocketChat.models.Messages = _.extend({}, RocketChat.models.Messages, Messages);
