+++ packages/rocketchat-lib/client/models/Roles.js
@@ -1,0 +1,4 @@
+import { Roles } from 'meteor/rocketchat:models';
+import _ from 'underscore';
+
+RocketChat.models.Roles = _.extend({}, RocketChat.models.Roles, Roles);
