+++ packages/rocketchat-lib/client/models/Rooms.js
@@ -1,0 +1,4 @@
+import { Rooms } from 'meteor/rocketchat:models';
+import _ from 'underscore';
+
+RocketChat.models.Rooms = _.extend({}, RocketChat.models.Rooms, Rooms);
