+++ packages/rocketchat-lib/client/models/Subscriptions.js
@@ -1,0 +1,4 @@
+import { Subscriptions } from 'meteor/rocketchat:models';
+import _ from 'underscore';
+
+RocketChat.models.Subscriptions = _.extend({}, RocketChat.models.Subscriptions, Subscriptions);
