+++ packages/rocketchat-lib/client/models/Users.js
@@ -1,0 +1,4 @@
+import { Users } from 'meteor/rocketchat:models';
+import _ from 'underscore';
+
+RocketChat.models.Users = _.extend({}, RocketChat.models.Users, Users);
