+++ packages/rocketchat-lib/lib/authorization.js
@@ -1,0 +1,1 @@
+RocketChat.authz = {};
