+++ packages/rocketchat-metrics/package.js
@@ -9,7 +9,7 @@ Package.onUse(function(api) {
 	api.use([
 		'ecmascript',
 		'rocketchat:settings',
-		'rocketchat:models',
+		'rocketchat:migrations',
 		'rocketchat:version',
 	]);
 	api.mainModule('server/index.js', 'server');
