+++ packages/rocketchat-models/client/models/CachedChannelList.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const CachedChannelList = new Mongo.Collection(null);
