+++ packages/rocketchat-models/client/models/CachedChatRoom.js
@@ -1,0 +1,3 @@
+import { CachedCollection } from './CachedCollection';
+
+export const CachedChatRoom = new CachedCollection({ name: 'rooms' });
