+++ packages/rocketchat-models/client/models/CachedChatSubscription.js
@@ -1,0 +1,3 @@
+import { CachedCollection } from './CachedCollection';
+
+export const CachedChatSubscription = new CachedCollection({ name: 'subscriptions' });
