+++ packages/rocketchat-models/client/models/CachedUserList.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const CachedUserList = new Mongo.Collection(null);
