+++ packages/rocketchat-models/client/models/ChatMessage.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const ChatMessage = new Mongo.Collection(null);
