+++ packages/rocketchat-models/client/models/ChatPermissions.js
@@ -1,0 +1,8 @@
+import { CachedCollection } from './CachedCollection';
+
+export const AuthzCachedCollection = new CachedCollection({
+	name: 'permissions',
+	eventType: 'onLogged',
+});
+
+export const ChatPermissions = AuthzCachedCollection.collection;
