+++ packages/rocketchat-models/client/models/ChatRoom.js
@@ -1,0 +1,3 @@
+import { CachedChatRoom } from './CachedChatRoom';
+
+export const ChatRoom = CachedChatRoom.collection;
