+++ packages/rocketchat-models/client/models/ChatSubscription.js
@@ -1,0 +1,3 @@
+import { CachedChatSubscription } from './CachedChatSubscription';
+
+export const ChatSubscription = CachedChatSubscription.collection;
