+++ packages/rocketchat-models/client/models/RoomRoles.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const RoomRoles = new Mongo.Collection(null);
