+++ packages/rocketchat-models/client/models/UserAndRoom.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const UserAndRoom = new Mongo.Collection(null);
