+++ packages/rocketchat-models/client/models/UserRoles.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const UserRoles = new Mongo.Collection(null);
